+++ Plugin.cs
@@ -15,7 +15,7 @@ namespace ServerSyncModTemplate;
 public partial class ServerSyncModTemplatePlugin : BaseUnityPlugin
 {
     internal const string ModName = "DiveIn";
-    internal const string ModVersion = "1.1.9";
+    internal const string ModVersion = "1.2.0";
     internal const string Author = "sighsorry";
     private const string ModGUID = $"{Author}.{ModName}";